+++ CHANGES.md
@@ -16,6 +16,12 @@ Fast and flexible PHP library for text tables.
 
 ## Changes
 
+* 2.7.0 (2023-10-15)
+  * Added `setNoDataLabel()` to override default `NO DATA` string shown when rendering empty table.
+  * Added `getContentTotalWidth()` method to the `TextTable`.
+  * Added option to control column title visibility.
+
+
 * 2.6.0 (2023-10-13)
   * Added `CompactRenderer`.
   * Renderers can now control rendering of top/bottom/separator rows.
@@ -24,6 +30,7 @@ Fast and flexible PHP library for text tables.
   * Added `Separator` row class, and `Table::addSeparator()` helper method.
   * Fixed `DuplicateColumnException` and `UnsupportedColumnAlignException` exceptions
   * Improved documentation.
+
 
 * 2.5.1 (2023-10-11)
   * Fixed `Column` c'tor overwriting inherited data/header align parameter.